backend/requirements.txt
pillow
django
djangorestframework
djoser
drf-yasg
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt